requirements.txt
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
